+++ CHANGELOG.md
@@ -119,7 +119,8 @@ console.log("txn1 information: ", receipts[2]);
 - Allow token to be empty.
 - Throw error when issue inner transactions in clear program. Fixed in [#667](https://github.com/scale-it/algo-builder/pull/667).
 - Parameters in `extract*` opcodes can greater than uint8. Fixed in [#666](https://github.com/scale-it/algo-builder/pull/666).
-- Restirct duplicate transaction in group transaction.
+- Wallet contructor come from a parameter walletURL(token, server, port)
+- Restirct duplicate transaction in group transaction. 
 
 ### Infrastructure